refactor(ff-filter): split builder.rs into builder/ submodule with inline tests
